project-01-chatbot/requirements.txt
gradio
python-dotenv
openai

